refactor: rename site_toolbox to sites_toolbox and migrate related functions
